+++ CHANGELOG.md
@@ -8,7 +8,15 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ### Changed
 
-- ...
+- (PHP SDK) Minimal PHP version now is `^7.1.3`
+- (PHP SDK) `StaticReference::getContent()` signature - now this method accepts two additional parameters - `bool $as_array = true, int $options = 0`
+
+### Added
+
+- Docker-based environment for development
+- Tests running using GitLab actions
+- (PHP SDK) `phpstan/phpstan` package for static code analyzer
+- (PHP SDK) Strict types and type-hinting
 
 ## v2.13.0